增加短信模板功能。

+++ guns-vip-main/src/main/java/cn/stylefeng/guns/modular/sms/model/params/SignnameParam.java
@@ -33,7 +33,7 @@ public class SignnameParam implements Serializable, BaseValidatingParam {
     /**
      * 企业ID
      */
-    private Integer entid;
+    private Long entid;
 
     /**
      * 黑名单类型
